remove recency constraint
Looks like the lambda is still running and scraping the posts, but it
isn't posting them for some reason. The guids are indeed not in the
'already posted' table so that leaves the other condition - the recency
condition.

Without going back through the logs too much, it's possible that AWS has
started updating the RSS feed more than 180 minutes after the post's
published time which would cause the lambda to skip all posts.

Napkin math suggests that at 100 posts per feed and a rate of 5 minutes,
just querying all 100 to dynamo every time will cost ~$0.21/mo which is
totally acceptable. So removing that condition